Support newer versions of gym
Version 0.9.6 of gym introduced several changes which broke
compatibility with most environments, including updates to the
MultiDiscrete interface and changes to the step, render, and observe
methods from private to public. This patch aims to add compatibility
with newer versions of gym by addressing the following:

* Update instantiation of MultiDiscrete action spaces to be non-negative
ranges with an inclusive lower-bound and exclusive upper-bound.
* Pad the actions being sent to the Controller class to ensure proper
ranges for joystick orientations are used. For example, the X- and
Y-axis ranges as instantiated with MultiDiscrete could be from [0, 256).
The controller, however, would expect numbers in the range of [-128,
127]. To account for this, the action supplied by gym should be wrapped
for the joystick axes before passing the action to the Controller class.
* Change all private methods required by gym to public methods (ie.
change `_step()` to `step()`).

Signed-Off-By: Robert Clark <robdclark@outlook.com>